refactor(#2681): extract abort taxonomy into obs_abort module; fmt
Split AbortReason/AbortContext/record_do_get_abort out of obs.rs into a sibling
obs_abort module (campsite rule, epic #1116) — obs.rs re-exports the stable
crate::obs::{...} path. Residual growth in service.rs/catalog.rs is intrinsic to
the feature on files already over the #1116 threshold (CQLITE_ALLOW_FILE_GROWTH).
